restructure qos
