+++ devkit_driver/devkit_driver/modules/__init__.py
@@ -2,12 +2,10 @@ from .bms_handler import BMSHandler
 from .bumper_handler import BumperHandler
 from .estop_handler import EStopHandler
 from .odom_handler import OdomHandler
-from .qos_safety import SAFETY_QOS
 from .robot_brain_handler import RobotBrainHandler
 from .twist_handler import TwistHandler
 
 __all__ = [
-    'SAFETY_QOS',
     'BMSHandler',
     'BumperHandler',
     'EStopHandler',
